@@ -4,7 +4,7 @@
 		"optimize-autoloader": true,
 		"classmap-authoritative": true,
 		"platform": {
-			"php": "7.3"
+			"php": "7.4"
 		}
 	},
 	"scripts": {
