@@ -1,10 +1,12 @@
 <?php
 
 declare(strict_types=1);
+
 /**
  * @copyright Copyright (c) 2016, ownCloud, Inc.
  *
  * @author Joas Schilling <coding@schilljs.com>
+ * @author Richard Steinmetz <richard@steinmetz.cloud>
  *
  * @license AGPL-3.0
  *
@@ -37,11 +39,8 @@ use Symfony\Component\Console\Formatter\OutputFormatterStyleInterface;
  * @package OCA\PreviewGenerator\Command
  */
 class TimestampFormatter implements OutputFormatterInterface {
-	/** @var IConfig */
-	protected $config;
-
-	/** @var OutputFormatterInterface */
-	protected $formatter;
+	protected IConfig $config;
+	protected OutputFormatterInterface $formatter;
 
 	/**
 	 * @param IConfig $config
@@ -57,7 +56,7 @@ class TimestampFormatter implements OutputFormatterInterface {
 	 *
 	 * @param bool $decorated Whether to decorate the messages or not
 	 */
-	public function setDecorated($decorated) {
+	public function setDecorated($decorated): void {
 		$this->formatter->setDecorated($decorated);
 	}
 
@@ -76,7 +75,7 @@ class TimestampFormatter implements OutputFormatterInterface {
 	 * @param string $name The style name
 	 * @param OutputFormatterStyleInterface $style The style instance
 	 */
-	public function setStyle($name, OutputFormatterStyleInterface $style) {
+	public function setStyle($name, OutputFormatterStyleInterface $style): void {
 		$this->formatter->setStyle($name, $style);
 	}
 
